pyproject.toml
[project]
name = "course-navigator-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "chromadb>=1.3.4",
    "langchain>=1.0.5",
    "openai>=2.7.2",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
]
